requirements.txt
fastapi==0.118.0
uvicorn[standard]==0.37.0
prometheus-fastapi-instrumentator==7.1.0
SQLAlchemy==2.0.43
psycopg[binary]==3.2.10
openai==1.99.1
python-dotenv==1.2.1
